Clean arch fix: move method to another repo and call it from use case
