+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2025-05-20T14:12:39.096Z\n"
-"PO-Revision-Date: 2025-05-20T14:12:39.096Z\n"
+"POT-Creation-Date: 2025-05-21T06:50:28.234Z\n"
+"PO-Revision-Date: 2025-05-21T06:50:28.234Z\n"
 
 msgid "Low"
 msgstr ""
